server/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
cryptography
python-multipart